Refactor contacts service to use separate folders
